requirements.txt
requests
uwsgi
python-decouple
Flask
